pyproject.toml
[tool.poetry]
name = "moxie"
version = "0.1.0"
description = ""
authors = ["Marcos Moyano <marcosmoyano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
markdown = "^3.7"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
